fix: address PR review : Add unique filename suffix generator, file cleanup for part-download
