+++ functions/upload-file/1.0/Cargo.toml
@@ -6,6 +6,8 @@ version = "0.1.0"
 [lib]
 crate-type = ["cdylib"]
 
+[workspace]
+
 [dependencies]
 wasmcloud-component = "0.2.0"
 anyhow = "1"
