+++ functions/upload-file/1.0/src/download.rs
@@ -1,23 +1,14 @@
 use anyhow::{Context, Result};
-use tracing::{debug, error};
-use waki::{header::HeaderName, Client, RequestBuilder, Response};
-
-use crate::bindings::wasi::filesystem::{
-    preopens::get_directories,
-    types::{DescriptorFlags, OpenFlags, PathFlags},
-};
-
-const CHUNK_SIZE: u64 = 65536; // 64kb
+use tracing::debug;
+use waki::{header::HeaderName, Client, RequestBuilder};
 
 pub fn download_and_stream_to_disk(
     client: &Client,
     url: &str,
     headers: Option<&[(String, String)]>,
-) -> Result<(u64, String, String)> {
+    file_name: &str,
+) -> Result<u64> {
     debug!("Downloading from: {}", url);
-
-    let (file_name, content_type) =
-        extract_file_info_from_url(url).context("Failed to extract file info from URL")?;
 
     let response = build_request(client, url, headers)?
         .send()
@@ -31,9 +22,19 @@ pub fn download_and_stream_to_disk(
         ));
     }
 
-    let file_size = stream_response_to_file(&response, &file_name)?;
+    let file_size = crate::fs::stream_response_to_file(&response, file_name)?;
 
-    Ok((file_size, file_name, content_type))
+    Ok(file_size)
+}
+
+pub fn make_unique_filename(filename: &str) -> String {
+    let random_bytes = crate::bindings::wasi::random::random::get_random_bytes(8);
+    let hex: String = random_bytes.iter().map(|b| format!("{b:02x}")).collect();
+
+    match filename.rsplit_once('.') {
+        Some((stem, ext)) => format!("{stem}_{hex}.{ext}"),
+        None => format!("{filename}_{hex}"),
+    }
 }
 
 fn build_request(
@@ -52,101 +53,6 @@ fn build_request(
     Ok(request)
 }
 
-fn stream_response_to_file(response: &Response, filename: &str) -> Result<u64> {
-    let preopens = get_directories();
-
-    if preopens.is_empty() {
-        error!("stream_response_to_file: No preopened directories available!");
-        return Err(anyhow::anyhow!("No preopened directories available"));
-    }
-
-    let (dir, _dir_name) = &preopens[0];
-
-    let file = dir
-        .open_at(
-            PathFlags::empty(),
-            filename,
-            OpenFlags::CREATE | OpenFlags::TRUNCATE,
-            DescriptorFlags::WRITE,
-        )
-        .map_err(|e| {
-            error!("stream_response_to_file: Failed to open file for writing: {e:?}");
-            anyhow::anyhow!("Failed to open file for writing: {e:?}")
-        })?;
-
-    let write_stream = file.write_via_stream(0).map_err(|e| {
-        error!("stream_response_to_file: Failed to get write stream: {e:?}");
-        anyhow::anyhow!("Failed to get write stream: {e:?}")
-    })?;
-
-    let mut total_bytes: u64 = 0;
-    let mut chunk_count: u64 = 0;
-
-    loop {
-        // stream the chunk, Question(Aditya): what should be the chunk size?
-        let chunk_result = response.chunk(CHUNK_SIZE);
-        match chunk_result {
-            Ok(Some(chunk)) if !chunk.is_empty() => {
-                chunk_count += 1;
-
-                if let Err(e) = write_chunk_to_stream(&write_stream, &chunk) {
-                    error!(
-                        "stream_response_to_file: Failed to write chunk {}: {}",
-                        chunk_count, e
-                    );
-                    let _ = dir.unlink_file_at(filename);
-                    return Err(e.context(format!("Failed to write chunk {}", chunk_count)));
-                }
-                total_bytes += chunk.len() as u64;
-            }
-            Ok(Some(_)) | Ok(None) => break,
-            Err(e) => {
-                error!("stream_response_to_file: Failed to read response chunk: {e:?}");
-                let _ = dir.unlink_file_at(filename);
-                return Err(anyhow::anyhow!("Failed to read response chunk: {e:?}"));
-            }
-        }
-    }
-
-    write_stream.flush().map_err(|e| {
-        error!("stream_response_to_file: Failed to flush write stream: {e:?}");
-        anyhow::anyhow!("Failed to flush write stream: {e:?}")
-    })?;
-    drop(write_stream);
-    drop(file);
-
-    Ok(total_bytes)
-}
-
-fn write_chunk_to_stream(
-    stream: &crate::bindings::wasi::io::streams::OutputStream,
-    chunk: &[u8],
-) -> Result<()> {
-    let mut offset = 0;
-    while offset < chunk.len() {
-        let to_write = &chunk[offset..];
-        match stream.check_write() {
-            Ok(0) => {
-                stream.subscribe().block();
-                continue;
-            }
-            Ok(available) => {
-                let write_size = std::cmp::min(available as usize, to_write.len());
-                stream.write(&to_write[..write_size]).map_err(|e| {
-                    error!("write_chunk_to_stream: Write failed: {e:?}");
-                    anyhow::anyhow!("Write failed: {e:?}")
-                })?;
-                offset += write_size;
-            }
-            Err(e) => {
-                error!("write_chunk_to_stream: check_write failed: {e:?}");
-                return Err(anyhow::anyhow!("check_write failed: {e:?}"));
-            }
-        }
-    }
-    Ok(())
-}
-
 pub fn extract_file_info_from_url(url: &str) -> Result<(String, String)> {
     let url_path = url
         .split('?')
@@ -160,7 +66,7 @@ pub fn extract_file_info_from_url(url: &str) -> Result<(String, String)> {
         .ok_or_else(|| anyhow::anyhow!("Could not extract filename from URL"))?;
 
     let filename = urlencoding::decode(encoded_filename)
-        .unwrap_or_else(|_| std::borrow::Cow::Borrowed(encoded_filename))
+        .unwrap_or(std::borrow::Cow::Borrowed(encoded_filename))
         .to_string();
 
     let content_type = mime_guess::from_path(&filename)
@@ -168,4 +74,49 @@ pub fn extract_file_info_from_url(url: &str) -> Result<(String, String)> {
         .to_string();
 
     Ok((filename, content_type))
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn test_extract_file_info_from_url_simple() {
+        let (filename, content_type) =
+            extract_file_info_from_url("https://example.com/files/document.pdf").unwrap();
+        assert_eq!(filename, "document.pdf");
+        assert_eq!(content_type, "application/pdf");
+    }
+
+    #[test]
+    fn test_extract_file_info_from_url_with_query() {
+        let (filename, content_type) =
+            extract_file_info_from_url("https://example.com/files/image.png?token=abc123").unwrap();
+        assert_eq!(filename, "image.png");
+        assert_eq!(content_type, "image/png");
+    }
+
+    #[test]
+    fn test_extract_file_info_from_url_encoded() {
+        let (filename, content_type) =
+            extract_file_info_from_url("https://example.com/files/my%20file.txt").unwrap();
+        assert_eq!(filename, "my file.txt");
+        assert_eq!(content_type, "text/plain");
+    }
+
+    #[test]
+    fn test_extract_file_info_from_url_unknown_extension() {
+        let (filename, content_type) =
+            extract_file_info_from_url("https://example.com/files/data.unknownext123").unwrap();
+        assert_eq!(filename, "data.unknownext123");
+        assert_eq!(content_type, "application/octet-stream");
+    }
+
+    #[test]
+    fn test_extract_filename_from_trailing_slash_link() {
+        let (filename, content_type) =
+            extract_file_info_from_url("https://example.com/files/data/somedir/test.pdf/").unwrap();
+        assert_eq!(filename, "test.pdf");
+        assert_eq!(content_type, "application/pdf");
+    }
 }
