+++ functions/upload-file/1.0/wasmcloud.lock
@@ -28,3 +28,12 @@ registry = "wasi.dev"
 requirement = "=0.1.0-draft"
 version = "0.1.0-draft"
 digest = "sha256:09621a45b12b0a9cddc798517f778aac0e5ae4bd234077b3d70758d6cf625580"
+
+[[packages]]
+name = "wasi:random"
+registry = "wasi.dev"
+
+[[packages.versions]]
+requirement = "=0.2.2"
+version = "0.2.2"
+digest = "sha256:143e47a99fe9486672b5f03491e8e610e413a646dc575059c6ebbb354d7c941a"
