+++ functions/upload-file/1.0/wit/world.wit
@@ -4,6 +4,7 @@ world upload {
     import wasi:filesystem/types@0.2.2;
     import wasi:filesystem/preopens@0.2.2;
     import wasi:http/outgoing-handler@0.2.2;
+    import wasi:random/random@0.2.2;
     import wasi:logging/logging@0.1.0-draft;
     import betty-blocks:data-api/data-api-utilities;
     import betty-blocks:types/types;
