--- test/rust/README.md
@@ -1,4 +1,0 @@
-## Note:(Aditya)
-
-I'm keeping the test file here till we figure out *where* to put or if we *even* want to have integration tests in this repo 
-If this test isn't supposed to be in this repo then lmk I'll discard it 👍.
